refactor(query): Extract mapping methods to planner/mapping.rs
Move document mapping builders (193 lines) from builder.rs:
- build_mapping() for Select operation mappings
- build_group_child_mapping() for _group virtual fields

builder.rs reduced from 4,621 to 4,398 lines.
Total extracted: views (206) + aggregates (305) + mapping (193) = 704 lines.
