+++ crates/query/src/planner/builder.rs
@@ -3652,228 +3652,6 @@ impl Planner {
         Ok((new_plan, mapping))
     }
 
-    /// Build the document mapping for a Select operation.
-    ///
-    /// IMPORTANT: _docID is ALWAYS placed at index 0 because Doc::doc_id() expects it there.
-    /// TypeJoinOne/TypeJoinMany use doc_id() to match related documents.
-    pub(super) fn build_mapping(
-        &self,
-        select: &Select,
-        collection: &CollectionVersion,
-    ) -> Result<DocumentMapping> {
-        let mut mapping = DocumentMapping::new();
-
-        // Track whether _docID was explicitly requested (for render_keys)
-        let mut doc_id_requested = false;
-        let mut doc_id_alias: Option<String> = None;
-
-        // Check if _docID is explicitly selected
-        for requestable in &select.fields {
-            if let Requestable::Field(field) = requestable {
-                if field.name == "_docID" {
-                    doc_id_requested = true;
-                    doc_id_alias = Some(field.output_name().to_string());
-                    break;
-                }
-            }
-        }
-
-        // ALWAYS add _docID at index 0 (required for Doc::doc_id() to work)
-        mapping.add(0, "_docID");
-        // Only add to render_keys if explicitly selected
-        if doc_id_requested {
-            mapping.add_render_key(0, doc_id_alias.as_deref().unwrap_or("_docID"));
-        }
-
-        // Add remaining requested fields (starting from index 1)
-        for requestable in &select.fields {
-            match requestable {
-                Requestable::Field(field) => {
-                    // Skip _docID (already handled at index 0)
-                    if field.name == "_docID" {
-                        continue;
-                    }
-                    // Handle _group specially - it's a virtual field for groupBy results
-                    if field.name == "_group" {
-                        let index = mapping.next_index();
-                        mapping.add(index, "_group");
-                        mapping.add_render_key(index, field.output_name());
-                        continue;
-                    }
-                    // Handle __typename for GraphQL introspection
-                    if field.name == "__typename" {
-                        mapping.set_type_name(&select.collection_name);
-                        let index = mapping.first_index_of_name("__typename").unwrap();
-                        mapping.add_render_key(index, field.output_name());
-                        continue;
-                    }
-                    // Handle _deleted (soft-delete status, populated at render time)
-                    if field.name == "_deleted" {
-                        let index = mapping.next_index();
-                        mapping.add(index, "_deleted");
-                        mapping.add_render_key(index, field.output_name());
-                        continue;
-                    }
-                    // Validate field exists in schema
-                    if collection.field_by_name(&field.name).is_none() {
-                        return Err(QueryError::unknown_field(&field.name));
-                    }
-                    let index = mapping.next_index();
-
-                    mapping.add(index, &field.name);
-                    mapping.add_render_key(index, field.output_name());
-                }
-                Requestable::Select(nested_select) => {
-                    // Handle _group specially - it's a virtual field for groupBy results
-                    if nested_select.field.name == "_group" {
-                        let index = mapping.next_index();
-                        mapping.add(index, "_group");
-                        mapping.add_render_key(index, nested_select.field.output_name());
-
-                        // Build child mapping for the _group nested fields
-                        let child_mapping =
-                            self.build_group_child_mapping(&nested_select, collection)?;
-                        mapping.set_child_at(index, child_mapping);
-                        continue;
-                    }
-                    // Nested select (relation) - add the field but don't recurse here
-                    // Child mapping will be built when applying joins
-                    let index = mapping.next_index();
-                    mapping.add(index, &nested_select.field.name);
-                    mapping.add_render_key(index, nested_select.field.output_name());
-                }
-                Requestable::Aggregate(agg) => {
-                    // For relation-based aggregates (e.g., _count(books: {})),
-                    // the aggregate operates on related documents.
-                    // We add the aggregate name to the mapping for the result field.
-                    let index = mapping.next_index();
-                    let name = agg.aggregate_type.as_str();
-                    mapping.add(index, name);
-                    // Use alias if provided, otherwise use the aggregate name
-                    mapping.add_render_key(index, agg.output_name());
-                }
-                Requestable::Similarity(sim) => {
-                    let index = mapping.next_index();
-                    mapping.add(index, "_similarity");
-                    mapping.add_render_key(index, sim.output_name());
-                }
-            }
-        }
-
-        // If no fields specified (besides _docID), add all collection fields
-        if mapping.next_index() == 1 {
-            // Only _docID was added; add all collection fields at schema indices
-            for (i, field) in collection.fields.iter().enumerate() {
-                if field.name != "_docID" {
-                    mapping.add(i, &field.name);
-                    mapping.add_render_key(i, &field.name);
-                } else if !doc_id_requested {
-                    // Add _docID to render_keys since we're returning all fields
-                    mapping.add_render_key(0, "_docID");
-                }
-            }
-        }
-
-        // Add fields referenced by the filter (needed for filter evaluation but not rendered)
-        if let Some(ref filter) = select.filter {
-            for field_name in filter.referenced_fields() {
-                if mapping.first_index_of_name(&field_name).is_none() {
-                    // Only add if field exists in collection schema
-                    if collection.field_by_name(&field_name).is_some() {
-                        let index = mapping.next_index();
-                        mapping.add(index, &field_name);
-                        // Don't add render_key - we don't want to output these fields
-                    }
-                }
-            }
-        }
-
-        Ok(mapping)
-    }
-
-    /// Build a child mapping for the _group virtual field.
-    /// This maps the nested fields within _group { ... } using the same indices
-    /// as the parent document, so field values can be extracted correctly.
-    fn build_group_child_mapping(
-        &self,
-        group_select: &Select,
-        collection: &CollectionVersion,
-    ) -> Result<DocumentMapping> {
-        let mut child_mapping = DocumentMapping::new();
-
-        // Build render_keys for each requested field, using schema indices
-        for requestable in &group_select.fields {
-            match requestable {
-                Requestable::Field(field) => {
-                    // Handle __typename for GraphQL introspection
-                    if field.name == "__typename" {
-                        child_mapping.set_type_name(&collection.name);
-                        let index = child_mapping.first_index_of_name("__typename").unwrap();
-                        child_mapping.add_render_key(index, field.output_name());
-                        continue;
-                    }
-
-                    // Handle _deleted (soft-delete status)
-                    if field.name == "_deleted" {
-                        let index = child_mapping.next_index();
-                        child_mapping.add(index, "_deleted");
-                        child_mapping.add_render_key(index, field.output_name());
-                        continue;
-                    }
-
-                    // Get the schema index for this field
-                    let schema_idx = if field.name == "_docID" {
-                        0
-                    } else {
-                        // Validate field exists in collection schema
-                        let schema_field = collection.field_by_name(&field.name);
-                        if schema_field.is_none() {
-                            return Err(QueryError::unknown_field(&field.name));
-                        }
-                        // Find the field's index in the collection
-                        collection
-                            .fields
-                            .iter()
-                            .position(|f| f.name == field.name)
-                            .unwrap_or(0)
-                    };
-
-                    child_mapping.add(schema_idx, &field.name);
-                    child_mapping.add_render_key(schema_idx, field.output_name());
-                }
-                Requestable::Select(nested_select) => {
-                    if nested_select.field.name == "_group" {
-                        // Nested _group within _group: recursively build child mapping
-                        let index = child_mapping.next_index();
-                        child_mapping.add(index, "_group");
-                        child_mapping.add_render_key(index, nested_select.field.output_name());
-
-                        let inner_child_mapping =
-                            self.build_group_child_mapping(nested_select, collection)?;
-                        child_mapping.set_child_at(index, inner_child_mapping);
-                    } else {
-                        // Relation field inside _group (e.g., published {...})
-                        // Add it to the child mapping so it can be rendered.
-                        // The actual index must match where TypeJoinMany will populate the data.
-                        // TypeJoinMany will use the parent mapping's index for this field.
-                        // We need to find or allocate an index for this relation field.
-                        let index = child_mapping.next_index();
-                        child_mapping.add(index, &nested_select.field.name);
-                        child_mapping.add_render_key(index, nested_select.field.output_name());
-                    }
-                }
-                Requestable::Aggregate(_) => {
-                    // Aggregates within _group are not currently supported
-                }
-                Requestable::Similarity(_) => {
-                    // Similarity within _group is not supported
-                }
-            }
-        }
-
-        Ok(child_mapping)
-    }
-
     /// Get a collection schema by name.
     pub fn collection(&self, name: &str) -> Option<&Arc<CollectionVersion>> {
         self.collections.get(name)
